Distinguish completion items from local/central provider
